Split parseContext out into own file
- Makes it easier to test / look at test coverage
